SignalingDimension.jl

Certify the classical simulation cost of black-box systems.

This Julia software package is the companion piece to Certifying the Classical Simulation Cost of a Quantum Channel.

Package Overview

SignalingDimension.SignalingDimension — Module

The advent of quantum technology requires new figures of merit to compare the performance of quantum and classical systems. The signaling dimension quantifies the classical simulation cost of black-box devices. That is, it specifies the minimum amount of noiseless, classical communication needed to exactly simulate the input-output statistics of the device in question. The signaling dimension is discussed in greater detail in the Background section of this documentation.

Features:

Tools to compute signaling correlations and polytopes (see Signaling Correlations and Signaling Polytopes).

A catalog of Bell inequalities for witnessing signaling dimension (see Bell Inequalities).

Methods to certify the signaling dimension of signaling systems (see Certifying Signaling Dimension).

Signaling Polytope Data

Our computed facets of the signaling polytope are found in the data/ directory. Please review the Computed Facets section for details regarding individual data sets.

Facet Verifications

This work provides scripts that verify the tightness of the Bell inequalities described in the Bell Inequalities section. These scripts and their results are found in the script/facet_verifications/ directory. Please refer to the Verification of Theoretical Facets section for more details.

Quick Start

Install Julia: https://julialang.org/downloads/

Add the SignalingDimension.jl package (run from julia prompt):

julia> using Pkg; Pkg.add("SignalingDimension")

To run the examples, add the BellScenario.jl and QBase.jl dependencies:

julia> using Pkg; Pkg.add("BellScenario")
julia> using Pkg; Pkg.add("QBase")

Contents

Signaling Correlations

Classical Channels

Quantum Channels

Signaling Polytopes

Vertices

Facets

Adjacency Decomposition

Signaling Dimension

Definition

Classical Simulation Cost

Significance

Certification

Bell Inequalities

Tight Bell Inequalities

Computed Facets

Theoretical Facets

Verification of Theoretical Facets

Certifying Signaling Dimension

Bounds

Quantum Channel Certification

Script Utilities

Running Scripts

Printing Test Results

Citing

To reference this work, see CITATION.bib.

Licensing

SignalingDimension.jl is released under the MIT License.

Acknowledgements

Development of SignalingDimension.jl was made possible by the advisory of Dr. Eric Chitambar and general support from the Physics Department at the University of Illinois Urbana-Champaign. Funding was provided by NSF Award 1914440.