Signaling Polytope: Bell Inequalities

The Bell inequalities (facets) of the signaling polytope bound the signaling correlations for $d$-dit classical communication. If a channel $\mathbf{P}\in\mathcal{P}^{n \to n'}$ violates a Bell inequality $\mathbf{F}\supset \mathcal{C}_d^{n \to n'}$, then $\mathbf{P}\notin\mathcal{C}_d^{n \to n'}$. Therefore, the signaling dimension of $\mathbf{P}$ is greater than $d$, that is, $\kappa(\mathbf{P}) > d$.

Theoretical Facets

We provide several tight Bell inequality which bound general signaling polytopes.

SignalingDimension.non_negativity_game — Function

non_negativity_game( num_outputs :: Int64, num_inputs :: Int64 ) :: BellGame

Constructs the non-negativity game for a channel with num_outputs and num_inputs.

A DomainError is thrown if num_outputs or num_inputs is not greater than 1.

SignalingDimension.success_game — Function

success_game( N :: Int64, d :: Int64 ) :: BellGame

Constructs the success game bound for the N-d-N polytope.

A DomainError is thrown if the following requirements aren't satisfied:

`N > 2

N > d > 1

SignalingDimension.ambiguous_game — Function

ambiguous_game( N :: Int64, d :: Int64 ) :: BellGame

Constructs the ambiguous game for the (N-1)-d-N polytope.

A DomainError is thrown if the inputs don't satisfy the following requirements:

N ≥ 4

(N - 2) ≥ d ≥ 2

SignalingDimension.error_game — Function

error_game( N :: Int64, d :: Int64, ε :: Int64 ) :: BellGame

Constructs the error game bound for the N-d-N polytope. Input ε sets the size of the anti-distinguishability matrix block.

A DomainError is thrown if the following requirements aren't satisfied:

N ≥ 4

(N - 2) ≥ d ≥ 2

(N - d + 1) ≥ ε ≥ 3

SignalingDimension.generalized_error_game — Function

generalized_error_game( N :: Int64, d :: Int64, k :: Int64 ) :: BellGame

Constructs the generalized error game for the specified parameters:

N is the number of outputs

d is the signaling dimension

k is the number of non-zero terms in each column

A DomainError is satisfied if the following requirements aren't satisfied:

(N - k) ≥ d ≥ 2

(N - 1) ≥ k ≥ 1

N ≥ 3

SignalingDimension.coarse_grained_input_ambiguous_game — Function

coarse_grained_input_ambiguous_game(
    num_outputs :: Int64,
    d :: Int64
) :: BellGame

Constructs a canonical form for a input coarse-grained ambiguous game.

A DomainError is thrown if the inputs don't satisfy the following requirements:

num_outputs ≥ 4

(num_outputs - 2) ≥ d ≥ 2

Affinely Independent Enumerations

To prove the tightness of the preceeding facet, a set of $n(n'-1)$ affinely independent vertices must be shown to satisfy $\gamma = \langle \mathbf{G}, \mathbf{V} \rangle$. These enumerations a demonstrated with the following methods.

SignalingDimension.aff_ind_success_game_strategies — Function

aff_ind_success_game_strategies( N :: Int64, d :: Int64 ) :: Vector{Matrix{Int64}}

Enumerates an affinely independent set of deterministic, stochastic, N x N rank-d matrices with trace equal to d. These matrices are all vertices of the success game facet found on the signaling polytope where N describes the number of inputs and d describes the signaling dimension of the communication channel.

A valid input requires N > 2 and N > d > 1.

SignalingDimension.aff_ind_non_negativity_game_strategies — Function

aff_ind_non_negativity_game_strategies(
    num_outputs :: Int64,
    num_inputs :: Int64
) :: Vector{Matrix{Int64}}

Enumerates an affinely independent set of deterministic, stochastic, num_outputs x num_inputs matrices that satisfy the non_negativity_game with equality.

A valid input requires N > 2 and N > d > 1.

SignalingDimension.aff_ind_ambiguous_game_strategies — Function

aff_ind_ambiguous_game_strategies(
    N :: Int64,
    d :: Int64
) :: Vector{Matrix{Int64}}

Enumerates an affinely independent set of deterministic, stochastic, N x (N-1) rank-d strategies which saturate the ambiguous_game. The existence of the enumeration proves that the ambiguous game is a facet of the signaling polytope

A DomainError is thrown if the inputs don't satisfy the following requirements:

N ≥ 4

(N - 2) ≥ d ≥ 2

SignalingDimension.aff_ind_coarse_grained_input_ambiguous_game_strategies — Function

aff_ind_coarse_grained_input_ambiguous_game_strategies(
    n :: Int64,
    d :: Int64
) :: Vector{Matrix{Int64}}

Enumerates an affinely independent set of deterministic, stochastic, n x n rank-d strategies which saturate the coarse_grained_input_ambiguous_game. The existence of the enumeration proves that the input coarse-graining lifting mechanism produces facets of the signaling polytope.

A DomainError is thrown if the inputs don't satisfy the following requirements:

n ≥ 4

(n - 2) ≥ d ≥ 2

SignalingDimension.aff_ind_error_game_strategies — Function

aff_ind_error_game_strategies(N :: Int64, d :: Int64, error_size :: Int64) :: Vector{Matrix{Int64}}

Enumerates an affinely independent set of vertices for the error game. The vertices are N x N, rank-d, column stochastic matrices.

Valid inputs are N > 4, N-1 > d > 1, and N-d+1 ≥ error_size ≥ 3.

SignalingDimension.aff_ind_generalized_error_game_strategies — Function

aff_ind_generalized_error_game_strategies(
    N :: Int64, d :: Int64, k :: Int64
) :: Vector{Matrix{Int64}}

Enumerates an affinely independent set of vertices for the generalized error game. The vertices are N x binomial(N,k), rank-d, column stochastic matrices. This function is restricted to games where N = d + k.

Valid inputs are N ≥ 4, d ≥ 2, and k ≥ 2.

Verification

Unit tests in [test/unit/affinely_independent_enumerations.jl] verify that each of these affinely independent enumerations scale across a wide range of scenarios.

Computed Facets

Using the adjacency decomposition technique, we've computed a broad range of signaling polytope facets. Computed facets of the signaling polytope are found in the data/ directory.

Quick Adjacency Decompositions

In the data/quick_adjacency_decomposition/ directory, the adjacency decomposition algorithm is used to find the generating facets of the signaling polytope. The polytope computation scripts are found in the script/quick_adjacency_decomposition/ directory. They are intended to run quickly on a laptop computer.

Data is provided in two formats:

.txt files are human readable

.ieq file format readable by BellScenario.jl.

Complete Polytopes

Verification

adjacency

complete computations